ata.device: Compiler delint (x86_64)

git-svn-id: https://svn.aros.org/svn/aros/trunk@39374 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ rom/devs/ata/lowlevel.c
@@ -1640,7 +1640,7 @@ BYTE ata_Identify(struct ata_Unit* unit)
             ULONG sec = unit->au_Capacity48;
 
             if (sec < unit->au_Capacity48)
-                sec = ~0ul;
+                sec = ~((ULONG)0);
 
             if (sec < unit->au_Capacity)
                 sec = unit->au_Capacity;
